pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "beatsaber-downloader"
version = "0.1.0"
description = "自动下载本地音乐对应的Beat Saber铺面"
authors = [{name = "BeatSaber Downloader Team"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mutagen>=1.47.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.25.0",
    "pydantic>=2.5.0",
    "loguru>=0.7.0",
    "PyYAML>=6.0",
    "aiofiles>=23.0.0",
    "orjson>=3.9.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
    "isort>=5.13.0",
    "bandit>=1.7.0",
]

[project.scripts]
beatsaber-downloader = "main:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]